requirements.txt
openai>=1.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
tqdm>=4.0.0
